requirements.txt
psycopg2-binary==2.9.6
psycopg==3.1.9
pyarrow==12.0.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
pgcopy==1.5.0
